+++ src/xrGame/WeaponMagazinedWGrenade.cpp
@@ -916,7 +916,7 @@ bool CWeaponMagazinedWGrenade::GetBriefInfo( II_BriefInfo& info )
 
 	GetSuitableAmmoTotal();
 
-	u32 at_size = m_bGrenadeMode ? m_ammoTypes2.size() : m_ammoTypes.size();
+	u32 at_size = m_bGrenadeMode ? (u32)m_ammoTypes2.size() : (u32)m_ammoTypes.size();
 	if(unlimited_ammo() || at_size == 0)
 	{
 		info.fmj_ammo._set("--");
