+++ src/xrGame/alife_simulator_base.cpp
@@ -273,7 +273,7 @@ void CALifeSimulatorBase::release	(CSE_Abstract *abstract, bool alife_query)
 	VERIFY							(object);
 
 	if (!object->children.empty()) {
-		u32							children_count = object->children.size();
+		u32							children_count = (u32)object->children.size();
 		u32							bytes = children_count*sizeof(ALife::_OBJECT_ID);
 		ALife::_OBJECT_ID			*children = (ALife::_OBJECT_ID*)_alloca(bytes);
 		CopyMemory					(children,&*object->children.begin(),bytes);
