+++ src/xrGame/moving_objects_dynamic.cpp
@@ -208,7 +208,7 @@ void moving_objects::generate_emitters			()
 	}
 #endif // DEBUG
 
-	u32							size = m_collision_emitters.size();
+	size_t size = m_collision_emitters.size();
 	m_collision_emitters.resize	(size + m_nearest_moving.size());
 
 	m_collision_emitters.erase	(
@@ -306,7 +306,7 @@ bool moving_objects::fill_collisions			(moving_object *object, const Fvector &ob
 	}
 
 	{
-		u32							collision_count = m_collisions.size();
+		size_t collision_count = m_collisions.size();
 		COLLISION_TIME				*collisions = (COLLISION_TIME*)_alloca(collision_count*sizeof(COLLISION_TIME));
 		std::copy					(m_collisions.begin(), m_collisions.end(), collisions);
 		COLLISION_TIME				*I_ = collisions;
@@ -458,7 +458,7 @@ void moving_objects::resolve_collisions			()
 
 	m_previous_collisions		= m_collisions;
 
-	u32							collidee_count = m_collisions.size()*2 + m_visited_emitters.size();
+	size_t collidee_count = m_collisions.size()*2 + m_visited_emitters.size();
 	moving_object				**collidees = (moving_object **)_alloca(collidee_count*sizeof(moving_object*));
 	{
 		moving_object			**J = collidees;
@@ -483,7 +483,7 @@ void moving_objects::resolve_collisions			()
 		collidee_count			= u32(std::unique(collidees, collidees + collidee_count) - collidees);
 	}
 
-	u32							decision_count = collidee_count;
+	size_t							decision_count = collidee_count;
 	decision					*decisions = (decision*)_alloca(decision_count*sizeof(decision));
 
 	{
